Added several properties to GameLogic. I now differentiate between a player's deck, a draw pile, and a heal pile. A player's deck is the combo of their draw and heal piles. The heal piles is where cards go after battle. They remain in the heal pile until the draw pile is empty, at which time the heal pile is shuffled, set to the draw pile, and cleared.
moveToBattleField() has been refined, and considers how many cards are already present on the battlefield, which determines the faceUp value of a card. This func also calls the checkDrawIsEmpty function after every time a card is moved from a draw pile.

A battle() func encapsulates the win/tie scenarios and the movement of cards between the piles, which inherently tracks the players' scores.

A checkDrawIsEmpty() func will make sure that the draw pile always has cards in it.