Changed the Suit enum to have rawType String. This way, when I test to see what the card arrays have in them, I return "spade" instead of the file path that represents that value. Made this change specifically for easier to read logging. This affected both CardModel and DeckModel.
I have been practicing functional programming and found an opportunity to replace some imperative code with some declarative code. I reduced a block of code that prints out the contents of the two player's decks from 7 lines to 2 lines. Notably, the output is a different and more desirable form: an array of tuples, each of which shows the suit and value property for that card.